OpenAeroVTOL Beta 52
Beta 51
 - Changed LCD_Display_Text() and do_menu_item() to index more than 256 items
 - Added very complicated auto-trim feature to allow independent mechanical trims for P1 and P2 to be set in flight. I-term neutral stick position resets after two seconds of no stick input if the Auto-trim feature is in use. 
 - S.Bus2 packets now supported for KK2.1+, thanks to OpenPilot code example.
 - Sped up EEPROM saves, removed LED flash. Tweaked menus to suit different feel. 
 - Code size improvements.
 - Auto-trim EEPROM save moved to throttle cut at end of flight. 
 - Dramatically reduced LCD startup time.
 - Improved code to stabilise FC when pressing calibrate button.

Beta 52
 - Now has "Quadcopter" compile option to assist rapid setup and testing.
 - Fixed MPU6050 filter setting being forgotten on power-up.
 - Restored missing logo. 
 - Sped up GLCD reset. 
 - Fixed orientation bugs.
 - Improved default AccZeroNormZ setting. 
 - Fixed throttle offset change when volumes both zero.
 - KK2 Mini compatibility tweaks. Change LCD initialisation and limit available contrast.
 - Added ADVANCED compile option to hide/expose the new features (Acro, Progressive P and Auto-trim)

+++ OpenAeroVTOL/OpenAeroVTOL_2_0/src/menu_driver.c
@@ -32,7 +32,7 @@ void print_menu_frame(uint8_t style);
 
 // Menu management
 void update_menu(uint8_t items, uint8_t start, uint8_t offset, uint8_t button, uint8_t* cursor, uint8_t* top, uint8_t* temp);
-void do_menu_item(uint8_t menuitem, int8_t *values, uint8_t mult, menu_range_t range, int8_t offset, uint8_t text_link, bool servo_enable, int16_t servo_number);
+void do_menu_item(uint16_t menuitem, int8_t *values, uint8_t mult, menu_range_t range, int8_t offset, uint8_t text_link, bool servo_enable, int16_t servo_number);
 void print_menu_items(uint8_t top, uint8_t start, int8_t values[], const unsigned char* menu_ranges, uint8_t rangetype, uint8_t MenuOffsets, const unsigned char* text_link, uint8_t cursor);
 
 // Misc
@@ -152,7 +152,7 @@ menu_range_t get_menu_range(const unsigned char* menu_ranges, uint8_t menuitem)
 // servo_number = Servo number to update
 //************************************************************
 
-void do_menu_item(uint8_t menuitem, int8_t *values, uint8_t mult, menu_range_t range, int8_t offset, uint8_t text_link, bool servo_enable, int16_t servo_number)
+void do_menu_item(uint16_t menuitem, int8_t *values, uint8_t mult, menu_range_t range, int8_t offset, uint8_t text_link, bool servo_enable, int16_t servo_number)
 {
 	mugui_size16_t size;
 	int16_t temp16;
